fix(ui): consolidate curve rendering code
